refactor: 都市データを constants/ へ移動し gitignore 回避を解消
frontend/src/data/ はリポジトリ直下の OSM 用 data/ を無視する
広域ルール data/ に巻き込まれ、!frontend/src/data/ の否定パターンが
必要だった。constants/ へ移すことで名前衝突が消え、否定行も不要になる。
